+++ src/bluesearch/entrypoint/database/parse.py
@@ -18,6 +18,7 @@
 from __future__ import annotations
 
 import argparse
+import gzip
 import json
 import logging
 import sys
@@ -140,7 +141,8 @@ def iter_parsers(input_type: str, input_path: Path) -> Iterator[ArticleParser]:
         yield PubMedXMLParser(input_path)
 
     elif input_type == "pubmed-xml-set":
-        articles = ElementTree.parse(str(input_path))
+        with gzip.open(input_path) as xml_stream:
+            articles = ElementTree.parse(xml_stream)
         for article in articles.iter("PubmedArticle"):
             yield PubMedXMLParser(article)