+++ tests/unit/entrypoint/database/test_parse.py
@@ -49,12 +49,6 @@ from bluesearch.entrypoint.database.parse import iter_parsers
             id="pubmed-xml",
         ),
         pytest.param(
-            "pubmed-xml-set",
-            "pubmed_articles.xml",
-            ["e9bb8ba085982a7cbb7d9ac2dbbafc7f", "49442b9ec575ae01f4934dfd79d03631"],
-            id="pubmed-xml-set",
-        ),
-        pytest.param(
             "tei-xml",
             "1411.7903v4.xml",
             ["73604b8751f2f4ecf63a5cefd042f6a3"],
@@ -71,6 +65,16 @@ from bluesearch.entrypoint.database.parse import iter_parsers
 def test_iter_parsers(input_type, path, article_uids):
     input_path = Path("tests/data/") / path
     parsers = iter_parsers(input_type, input_path)
+    for parser, uid in zip(parsers, article_uids):
+        assert parser.uid == uid
+
+
+def test_iter_parsers_pubmed_xml_set(pubmed_xml_gz_path):
+    parsers = iter_parsers("pubmed-xml-set", pubmed_xml_gz_path)
+    article_uids = [
+        "e9bb8ba085982a7cbb7d9ac2dbbafc7f",
+        "49442b9ec575ae01f4934dfd79d03631",
+    ]
     for parser, uid in zip(parsers, article_uids):
         assert parser.uid == uid
 
@@ -153,10 +157,10 @@ def test_cord19_json(jsons_path, tmp_path, caplog):
     assert "Argument 'input_path'" in caplog.text
 
 
-def test_pubmed_xml_set(tmp_path):
-    input_path = "tests/data/pubmed_articles.xml"
-    main(["parse", "pubmed-xml-set", input_path, str(tmp_path)])
-    files = sorted(tmp_path.iterdir())
+def test_pubmed_xml_set(pubmed_xml_gz_path, tmp_path):
+    output_dir = tmp_path / "test"
+    main(["parse", "pubmed-xml-set", str(pubmed_xml_gz_path), str(output_dir)])
+    files = sorted(output_dir.iterdir())
     assert len(files) == 2
 
     uids = ["49442b9ec575ae01f4934dfd79d03631", "e9bb8ba085982a7cbb7d9ac2dbbafc7f"]